extract findRoot function
